pyproject.toml
[tool.poetry]
name = "weather-bot"
version = "0.1.0"
description = ""
authors = ["alina1412 <alina1412>"]
readme = "README.md"
packages = [{include = "weather_bot"}]

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "^2.23.1"
python-dotenv = "^0.21.0"
httpx = "^0.23.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
isort = "^5.10.1"
black = "^22.10.0"
pytest-asyncio = "^0.20.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
markers = [
    "my: for debugging purposes",
]

asyncio_mode = "auto"